Make AppLayout the Root Component
Refactorbus
